Some ubsan fixes.
alignment_macros.h:
- New macros IS_ALIGNED_OFFSET() and IS_ALIGNED_PTR() to check if a
  given offset or pointer has the specified alignment.

common.h:
- Add RP_ALIGNED(n) for gcc's aligned attribute.

WiiUAncast::isRomSupported_static():
- data0_offset must be 32-bit aligned. Otherwise, if something is
  misdetected, we might end up doing an unaligned 32-bit read
  within a WiiU_Ancast_Header_SigCommon_t*.

src/libromdata/Console/WiiUAncast.cpp:355:19: runtime error: member access within misaligned address 0x7ffeeb9f84fb for type 'const WiiU_Ancast_Header_SigCommon_t' (aka 'const struct _WiiU_Ancast_Header_SigCommon_t'), which requires 4 byte alignment
0x7ffeeb9f84fb: note: pointer points here
 00  00 00 00 00 00 00 00 00  00 00 00 00 00 00 00 00  00 00 00 00 00 00 00 00  e6 ff ff ff 00 00 00
              ^

XboxXPRPrivate::loadXboxXPR0Image():
- Cast width and height to off64_t to avoid an integer overflow.
- Divide bpp by 8 first, before multiplying the rest.
src/librptexture/fileformat/XboxXPR.cpp:409:46: runtime error: signed integer overflow: 1073741824 * 16 cannot be represented in type 'int'

exe_pe_structs.h:
- Use 32-bit alignment for some structs that contain 64-bit values,
  since Windows doesn't always use 64-bit alignment.

src/libromdata/Other/EXE_PE.cpp:682:24: runtime error: reference binding to misaligned address 0x5628a86bfdbc for type 'const IMAGE_OPTIONAL_HEADER64' (aka 'const struct _IMAGE_OPTIONAL_HEADER64'), which requires 8 byte alignment
0x5628a86bfdbc: note: pointer points here
  e0 00 02 03 0b 02 02 0a  00 8a 00 00 00 7c 00 00  00 00 00 00 04 88 00 00  00 10 00 00 00 a0 00 00
              ^
src/libromdata/Other/EXE_PE.cpp:683:25: runtime error: load of misaligned address 0x5628a86bfdec for type 'const uint16_t' (aka 'const unsigned short'), which requires 8 byte alignment
0x5628a86bfdec: note: pointer points here
  04 00 00 00 04 00 00 00  00 00 00 00 00 30 01 00  00 04 00 00 3d 2a 01 00  02 00 00 00 00 00 10 00
              ^
src/libromdata/Other/EXE_PE.cpp:688:13: runtime error: member access within misaligned address 0x5628a86bfe9c for type 'const IMAGE_DATA_DIRECTORY' (aka 'const struct _IMAGE_DATA_DIRECTORY'), which requires 8 byte alignment
0x5628a86bfe9c: note: pointer points here
  00 00 00 00 2e 74 65 78  74 00 00 00 0c 88 00 00  00 10 00 00 63 67 69 3f  02 00 00 00 00 00 00 00
              ^

+++ src/librptexture/fileformat/XboxXPR.cpp
@@ -2,7 +2,7 @@
  * ROM Properties Page shell extension. (librptexture)                     *
  * XboxXPR.cpp: Microsoft Xbox XPR0 texture reader.                        *
  *                                                                         *
- * Copyright (c) 2019-2025 by David Korth.                                 *
+ * Copyright (c) 2019-2026 by David Korth.                                 *
  * SPDX-License-Identifier: GPL-2.0-or-later                               *
  ***************************************************************************/
 
@@ -406,7 +406,7 @@ rp_image_const_ptr XboxXPRPrivate::loadXboxXPR0Image(void)
 
 	// Determine the expected size based on the pixel format.
 	const auto &mode = xpr_mode_tbl[xpr0Header.pixel_format];
-	const size_t expected_size = width * height * mode.bpp / 8U;
+	const off64_t expected_size = static_cast<off64_t>(width) * static_cast<off64_t>(height) * mode.bpp / 8U;
 
 	if (expected_size > file_sz - data_offset) {
 		// File is too small.
@@ -415,8 +415,8 @@ rp_image_const_ptr XboxXPRPrivate::loadXboxXPR0Image(void)
 
 	// Read the image data.
 	auto buf = aligned_uptr<uint8_t>(16, expected_size);
-	size_t size = file->seekAndRead(data_offset, buf.get(), expected_size);
-	if (size != expected_size) {
+	size_t size = file->seekAndRead(data_offset, buf.get(), static_cast<size_t>(expected_size));
+	if (size != static_cast<size_t>(expected_size)) {
 		// Seek and/or read error.
 		return nullptr;
 	}
